Added tenant IDs to client API and messages
Summary:
Added a tenant ID to the client API and associated messages.

In the end I went with a 16-byte ID instead of a string. I figured that people are going to have to register an SLA with us anyway, so we might as well just provide an ID in the process and save several bytes per message instead of wasting bytes for a small amount of client convenience.

Also need to add some encoding functions for 16-bit as those were missing.

Test Plan: Tests added.

Reviewers: dhruba

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D20313