Log Router
Summary: Adds a utility class for use by the Pilot and Control Towers to find out what log a topic is mapped to. Handles the different retention classes correctly, and uses the fast Jump Consistent Hashing for the mapping. The only input is the total number of logs, which we will provide in a config.

Test Plan: Unit test added.

Reviewers: dhruba

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D20781