Create a generic MessageLoop and an EventLoop.
Summary:
The EventLoop does not really understand rocketspeed messages.
It knows the length of each message and is responsinble to
extract/encode/decode a message to/from a socket.

The MsgLoop understands the semantics of a message contents
and can invoke application-specified callbacks for each of the MsgType.
If the application does not specify a callback for a
particular MsgType, then those messages will be silenty dropped.

Test Plan: make check

Reviewers: pja

Reviewed By: pja

Differential Revision: https://reviews.facebook.net/D20973