LogDevice interface
Summary:
Work-in-progress interface for LogDevice. Submitting so that we can try to link it. Currently it will compile, but not link with LogDevice.

The Makefile prunes src/logdevice/storage.cc in build_detect_platform. Just remove the PRUNE_LOGDEVICE to try and compile/link it.

Test Plan: Work in progress. Once this is linking, I will add tests.

Reviewers: dhruba

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D20931