ControlTowerRouter
Summary:
Adds ControlTowerRouter, which maps log IDs to control tower URLs. It is just a thin wrapper around ConsistentHash<LogID, URL>, but will allow us to change the implementation as more functionality is needed, or for tuning the mapping.

To find a control tower for a particular topic, you will first use LogRouter to map the topic to a log ID, then ControlTowerRouter to map the log ID to a control tower.

Some additional functionality will be needed (e.g. a gossip protocol) to synchronize the mapping among CoPilots. This is not implemented here.

Test Plan: Unit test included

Reviewers: dhruba

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D20985