Mock LogDevice implementation using locked files
Summary:
Adds an implementation of the LogDevice interface that we can use for testing.

The implementation defines all functions required by the fbcode LogDevice headers, so we can easily link with either this or the real LogDevice. By default, the Makefile links with the mock implementation from this diff, and the real LogDevice (which currently doesn't link) can be linked against by exporting USE_LOGDEVICE=1.

The mock logdevice works by writing to files named "logdevice_<logid>.log". The log file format is a very simple and inefficient format that just serializes each record one after the other. Readers have to read the entire log to process new records (but use the last modified time on the file first). Writing new records is just an append to the end of the file.

This implementation should be suitable for testing purposes until we have a real LogDevice instance running. If needed we could cache the last read file offset in the readers to avoid processing entire logs.

Test Plan: Unit tests included

Reviewers: dhruba

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D21297