Implement a Ping message.
Summary: Implement a Ping message.

Test Plan:
make check
The ping unit test does not run yet, but this patch modifies quite a few
data structures, and fixes a bunch of elementary bugs. so it is better to
check it in (so that I can unblock PeterA) before I continue to fix the
remainder of the unit test.

Task ID: #

Blame Rev:

Reviewers: pja

Reviewed By: pja

Differential Revision: https://reviews.facebook.net/D22017