Event loop fixes
Summary:
Fixes for event loop shutdown and message decoding. Change over to use evconnlistener to manage connection setup.

The shutdown works by creating an eventfd listener event, which we can write to, to signal to the event loop that it should stop in a thread safe manner. The EventLoop destructor will send that message and wait in a loop for the event loop to process it.

Test Plan: make check

Reviewers: dhruba

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D22449