Fix bug with connection pooling.
Summary:
The connection pool code asserts if concurrent threads
tries to concurrently send message using MsgClient.Send().

Test Plan:
controlmessages_test now creates as many ControlRooms as there
are cpus on the machine. My machine has 24 cores, so the number
of ControlTower threads concurrently trying to send a response
back to the test is 24. This triggers this assertion.

Reviewers: pja

Reviewed By: pja

Differential Revision: https://reviews.facebook.net/D22719