Process a subscription/unsubscription message using LogStorage.
Summary:
Process a subscription/unsubscription message using LogStorage.
A subscription message can cause a new Log to be opened for reading,
or cause a seek to a new position in an already open log.
Similarly, an unsubscription message can cause an already open
log to be closed (if there are no active subscriptions on any
other topics in that log).

This is for early review, I will be removing the lint errors in a day.

Test Plan: make check

Reviewers: pja

Reviewed By: pja

Differential Revision: https://reviews.facebook.net/D23385