Adding Pilot worker loops
Summary:
Added pilot workers and a number of related changes:

* Remove the type from Command.
* Fix up several TARGETS files.
* Allow event loops without listener ports.
* Disabled libevent debugging in the pilotmessages_test because it isn't thread safe (now we have multiple loops)
* Added AppendAsync to the public API for non-blocking appends.

Note: the SimpleLogStorage and mock LogDevice implementations still block for AppendAsync. Only the real LogDevice will be non-blocking. The reason for this is because it is non-trivial to make them non-blocking. We can't just spawn a new thread for each AppendAsync since we'll run out of threads, so there needs to be an event loop or thread pool. This is non-trivial.

Test Plan: ./pilot &\sleep 1 && ./producer; pkill pilot

Reviewers: dhruba

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D23517