Change control tower router to map to HostIds instead of URLs
Summary:
ControlTowerRouter now maps log IDs to HostId*s instead of URLs, and maintains a set of HostIDs so that we can just store a pointer in the consistent hash ring.

Also updated the hasher to be able to hash multiple objects at once (the hostname + port tuple).

Test Plan: make check

Reviewers: dhruba

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D23643