Copilot first pass implementation
Summary: First pass implementation of the copilot. The simple test shows that it successfully can receive and respond to metadata requests, but I haven't test the logic thoroughly yet, or the data forwarding from log device. Just want to get this in as it's starting to become a large change.

Test Plan: make check

Reviewers: dhruba

Reviewed By: dhruba

Differential Revision: https://reviews.facebook.net/D24051