partially fixed ConsistentHash handling of equal hashes
Summary: Just stumbled on these two bugs while reading code. Here's a test for it and a fix. The test freezes forever without the fix. There's still a problem that the order of slots with the same hash depends on the order in which the slots were added, so in the very unlikely cases the hash function depends on the order of adding slots. Probably it would be better to forbid collisions altogether and just crash or something if it's discovered. As an alternative, we can make slots comparable (not just their hashes) and order them consistently.

Test Plan: make clean && make check

Reviewers: dhruba, pja

Reviewed By: pja

Differential Revision: https://reviews.facebook.net/D24381