Implement logging macros
Summary:
Create macros `LOG`, `LOG_DEBUG`, `LOG_INFO`, `LOG_WARN`, `LOG_ERROR`
and `LOG_FATAL` that work similarly to the `Log`, `Debug`, `Info`,
`Warn`, `Error` and `Fatal` functions, but do not evaluate their format
arguments if the specified log level is lower that the current minimal log
level.

Test Plan:
Run a compiled executable and check that the log messages are being
written correctly. Try changing the log level and check that only the
messages on that or higher level are being logged.

Reviewers: dhruba, pja

Reviewed By: pja

Differential Revision: https://reviews.facebook.net/D24837