Replace unix pipe with folly::ProducerConsumerQueue
Summary:
Replace the unix pipe that was used to pass commands from multiple producers to
a single consumer with
`folly::ProducerConsumerQueue` and two `eventfd`s to notify the consumer of new
commands and the producers of free spots in the queue. Because there can be
multiple producers, the write to the queue has to be guarded with a mutex.

Test Plan: make check

Reviewers: dhruba, pja

Reviewed By: pja

Differential Revision: https://reviews.facebook.net/D25011