There can be multiple outstanding socket connections to the same remote host.
Summary:
There can be more than one existing socket connections between one
msg loop and another. Let's assume that there is a existing connection
between Loop1 and Loop2. Now, if Loop1 encounters an error while
writing to this socket, it closes the connection immediately.
But Loop2 might not receive the socket-close event immediately.
Now Loop1 re-connects to Loop2 via a new connect request.

Loop2 will accept this connect request and create a new SocketEvent
inspite of the fact that there already exists an earlier SocketEvent
that is connected to Loop1 (but that earlier one is not functional
any more because Loop1 had closed that socket).

The above condition was causing an assert to trigger when connections
were being closed while running rocketbench. This patch prevents this
assert from occuring because we now store multiple SocketEvents for
a single remote host.

In fact, this patch will be beneficial in future when we might want
a pool of valid and functional connections between any two msg-loops
to alleviate transport bottlenecks.

Test Plan:
rocketbench
make check

Reviewers: pja

Reviewed By: pja

Differential Revision: https://reviews.facebook.net/D27699