Multiple event loops per message loop
Summary:
Main contribution of this patch is to add multiple EventLoops per MsgLoop. This allows greater concurrency on the server and client.

The MsgLoop now creates a configurable number of "worker" event loops, running on separate threads. They all call into the same callbacks, so the `Process*` functions are now potentially called from multiple threads. This turned out to only be an issue on the client, but can be solved by introducing some thread local data structures (not done in this patch, currently uses a mutex).

Only the first event loop does the accept. On accept, the MsgLoop sends the connection to another event loop for load balancing.

I have removed PilotWorker. The pilot now calls directly into LogDevice from the EventLoop thread.

For the Copilot and ControlTower, logs are mapped to the event loop workers the same way they are mapped to the rooms.

Made a small change to the EventLoop where it tries to write to a socket immediately instead of Enqueueing a write request (saves constantly adding/removing libevent events).

Changed RocketBench to use Statistics instead of its manual implementation.

Maybe we can discuss this change when you are back in tomorrow.

Test Plan: make check

Reviewers: dhruba

Reviewed By: dhruba

Subscribers: stupaq-fb, rocketspeed

Differential Revision: https://reviews.facebook.net/D29481