Fix rooms forwarding commands to the wrong tower loop
Summary:
Before this diff, when the copilot connected to the control tower, it would write to one control tower worker, but then the control tower would reply on a different worker. This patch ensures that the control tower rooms know which event loop to talk to for each subscriber copilot.

Also added some better logging which helped me debug this issue.

Test Plan: make check

Reviewers: rocketspeed, dhruba, stupaq-fb

Reviewed By: dhruba, stupaq-fb

Subscribers: stupaq-fb

Differential Revision: https://reviews.facebook.net/D29727