Fix for client sequence number filter
Summary: If a client received the same message with same seqno then it would process it twice. The topic map indicates the last read message, so I changed the filter relation to <= and used s-1 in the subscription response so that we get the first message at seqno s.

Test Plan: make check

Reviewers: rocketspeed, stupaq-fb, ranji42

Reviewed By: stupaq-fb, ranji42

Differential Revision: https://reviews.facebook.net/D29949