Just round-robin the load balancing for accepts
Summary: Disabling the dynamic load balancer because it isn't balancing load very well :-)  Since the accepts are kind of deferred now (they are written to the command queue then processed at a later time), the GetLoadFactor heuristic doesn't work immediately. This means that many connection can end up going to the same thread if the accepts all come in at the same time (as they do in the benchmark), which can ruin performance (as it is). I've disabling this so that we fall back on the round robin approach for now.

Test Plan: make check

Reviewers: rocketspeed, ranji42

Reviewed By: ranji42

Differential Revision: https://reviews.facebook.net/D30339