Factor out mutex+queue into multi-producer queue
Summary:
The WorkerLoop and EventLoop both using a ProducerConsuerQueue guarded with a mutex on the write side. This change factors this out into a MultiProducerQueue, which is the same thing in a separate class. The idea is that a bootcamper (or anyone) can change this into a real lockless multi producer queue at some point without disrupting other code.

I also noticed that all our WorkerLoops are multi-producer, so I just removed the option instead of adding another template parameter for the queue.

Test Plan: make check

Reviewers: rocketspeed, stupaq-fb

Reviewed By: stupaq-fb

Subscribers: stupaq-fb, dhruba

Differential Revision: https://reviews.facebook.net/D30489