Update HostNumber and WorkerId atomically.
Summary:
Update HostNumber and WorkerId atomically.
In the ComntrolTower, the HostMap maps a ClientId to a HostNumber.
The Control Tower also maintains a mapping from the HostNumber to
a WorkerId. The pre-existing code was updating the HostMap first
before updating the WorkerId-map. This can cause bugs because the
Lookup patch always assume that if a valid HostNumber exists, then
the WorkerId-map is correct.
The fix is to update the WorkerId-map before updating the HostMap.

Test Plan:
Run ./controlmessages_test 100 times. The assertion was repeatedly
reproduced prior to this patch.

Reviewers: pja

Reviewed By: pja

Subscribers: rocketspeed

Differential Revision: https://reviews.facebook.net/D30897