Every message sent can specify whether it is a request or response message.
Summary:
Every message sent can specify whether it is a request or response message.
The underlying transport can use this property to optimize
certian behaviour, e.g. if it is a response message then a
connection-based-transport can afford to not setup
new connections to the remote side.

Test Plan: make check

Reviewers: stupaq-fb, pja

Reviewed By: pja

Subscribers: rocketspeed

Differential Revision: https://reviews.facebook.net/D31269