Refactor server to allow pluggable storage implementation
Summary:
The diff has made a mess of this. The change is that instead of server/main.cc creating the log storage as well as all the rocketspeed services, it now just creates the log storage and then passes that to server.cc:Run, which does all the RocketSpeed specific stuff.

The idea is that in fbcode we can have a different main that passes in a different log storage, and share the code for setting up all the RocketSpeed stuff.

Test Plan:
make check # in github
fbmake opt && rocketspeed/github/src/tools/rocketbench/benchmark.sh --remote --deploy consume  # in fbcode

Reviewers: rocketspeed, stupaq-fb

Reviewed By: stupaq-fb

Differential Revision: https://reviews.facebook.net/D32385