Fix lack of output when asserting from different thread
Summary:
Test harness didn't produce callstack when asserting from a thread other than the main thread. This is because we throw an exception which is only caught by the main thread (in testharness.cc). If the exception isn't caught then the terminator doesn't get called and the error output never reaches stderr.

I've changed it so that the error is printed before firing the exception, rather than in the destructor.

Test Plan: make check

Reviewers: rocketspeed, stupaq-fb

Reviewed By: stupaq-fb

Differential Revision: https://reviews.facebook.net/D33123