Compile entire RocketSpeed on MacOS.
Summary:
All servers compile but the binaries do not link(yet) because of
a lack of libevent port.  libevent will be ported in a followup patch.

The visibility pragma for gcc compiler on macos/ios is not available
(but one can mark each  individual object as visible/notvisible).

Test Plan: make check on linux, macos and android.

Reviewers: pja, stupaq-fb

Reviewed By: stupaq-fb

Subscribers: rocketspeed

Differential Revision: https://reviews.facebook.net/D33945