Gather operation for MsgLoop
Summary: Adds a function to MsgLoop that allows us to do some computation on each event loop thread and then asynchronously gather the results. This is useful for testing the state of all event loops in tests, and will also be useful for gathering statistics in a thread-safe, non-locking way. I need this to properly test that goodbye messages are working correctly.

Test Plan: make check

Reviewers: rocketspeed, dhruba

Reviewed By: dhruba

Subscribers: dhruba

Differential Revision: https://reviews.facebook.net/D34227