[Rocketspeed] Modify client trim API and add gaps to be processed to reader.
Summary: Currently, the AsyncReader implementation does not process gaps, but instead skips them. In order to test propagation of gaps to the client (currently being worked on by @pja), the reader implementation must make the client aware of when gaps are found in the logs due to trimming. This change was made, as well as a change to the Client::Trim definition, which now uses SequenceNumber instead of an age. Finally, several integration tests were added to ensure that the changes to trimming and reading led to gaps being found by the reader and the appropriate callback called.

Test Plan: Added some integration tests.

Reviewers: rocketspeed, pja

Reviewed By: pja

Subscribers: pja

Differential Revision: https://reviews.facebook.net/D35775