Graceful shutdown
Summary:
Gracefully shutdown LogDevice and all services in such a way that callbacks don't hold dangling pointers.

This is a little complex since the only way to stop LogDevice firing callbacks is to delete all readers, and delete the client. We can't do that while the pilot or control tower are running though, so we have to first stop all our message loops to ensure we don't touch the client. Also, they are ref-counted, so we need to carefully ensure all refs are dropped (checked with asserts).

Test Plan: fbmake -r --sanitize=address rocketspeed/github/src && fbmake runtests

Reviewers: rocketspeed, stupaq-fb

Reviewed By: stupaq-fb

Subscribers: dhruba

Differential Revision: https://reviews.facebook.net/D37383