New timeout list to handle item expiry
Summary: ordered list of items based on the time they were inserted. If the same item is inserted again, the timestamp is updated and the item pushed to the back of the list. At any point, the items can be checked for an elapsed/timeout time and evicted from the list.

Test Plan: make check

Reviewers: rocketspeed, pja

Reviewed By: pja

Subscribers: pja, dhruba

Differential Revision: https://reviews.facebook.net/D37587