Switched stats to nonatomic and implemented GetStatistics in terms of MsgLoop::Gather
Summary: Stats is currently implemented using atomic memory operations making basic updates expensive. Switched to use non-atomic memory operations, and changed GetStatistics to use the MsgLoop::Gather pattern, and renamed to GetStatisticsSync to indicate the operation is more costly.

Test Plan: Planning on writing a unit test. Wanted to get this up to make sure this is the right direction given the task.

Reviewers: rocketspeed, stupaq-fb

Reviewed By: stupaq-fb

Subscribers: rocketspeed

Differential Revision: https://reviews.facebook.net/D36063