Remove metadata responses
Summary:
Stop sending responses for metadata messages. Client will only receive metadata when the server actively refuses a subscription (e.g. due to access control), but we don't do this for any existing use case.

When subscribing to 0, instead of getting a metadata messages telling you the actual sequence number you are subscribed to, the first message you receive will have the previous seqno set to 0 (so you can tell it apart from regular messages, and it keeps MessageType::mDeliver temporally commutative).

The 0-seqno logic is now in the TopicTailer instead of ControlRoom (since the TopicTailer needs to prepare the 0-seqno MessageData).

Fixed some bugs in the TopicTailer where the previous seqno tracking was incorrect after rewinds and other situations. The TopicTailer also now keeps track of a lower-bound estimate on the tail seqno, which we may be able to use for optimizations on sub(0) later on.

Test Plan: make check

Reviewers: rocketspeed, stupaq-fb

Reviewed By: stupaq-fb

Subscribers: stupaq-fb, dhruba

Differential Revision: https://reviews.facebook.net/D38325