Preliminary support for multiple readers.
Summary:
This adds support for multiple readers on a single log from a TopicTailer, but without the logic for selecting a good reader for a new subscription and merging readers when they catch up (just trying to keep the diffs small for easy review).

This was mostly straightforward with one notable change:

* Before, the LogReaders ref-counted the number of topic and log subcriptions we had.
* Now, all LogReaders forget about a topic when there are no longer any subscriptions on that topic (at all), and forget about logs when there are no topics subscribed on that log.

This is an all-round improvement I think: no more ref counting logic, and no need to keep track of which reader a hostnum+topic "belongs" to (for ref counting).

Test Plan: make check all

Reviewers: rocketspeed, stupaq-fb

Reviewed By: stupaq-fb

Subscribers: stupaq-fb, dhruba

Differential Revision: https://reviews.facebook.net/D39465