More intelligent selection of reader channels for new subscriptions
Summary:
With 1 reader per room, we use existing algorithm: always rewind if we have to.

With > 1 reader (we default to 2), we use the following algorihtm:

* If you subscribe ahead of a reader, we just use that reader.
* If you subscribe and there is an inactive reader, we use that reader.
* If there's no available readers, we assign you to a "pending reader" (not reading any logs)
* Once two readers reach the same seqno, we merge them, and the new available once picks up the subscriptions from pending reader.

This way the tail should always advance, and lagging subscriptions will be periodically picked up.

A better algorithm may dynamically create readers, or rewind based on some period of forward progress, but this should be good for now.

Test Plan:
make check all  # with various number of readers per room.

Also ran benchmarks.

Reviewers: rocketspeed, stupaq-fb

Reviewed By: stupaq-fb

Subscribers: stupaq-fb, eduardosuarez, dhruba

Differential Revision: https://reviews.facebook.net/D39555