Tower handle goodbyes
Summary:
Remove all suscriptions for a host on copilot goodye.

I don't like the approach, but perhaps the CPU/memory trade off is ok. I'd like opinions. Basically, on good bye, every room will search all subscriptions for the host and remove it. If we have, say, 10M subscriptions per control tower, then that's 250k per room. Iterating that should be pretty quick.

The alternative approach is to somehow store the list of topics that a host is subscribed to so that the work is O(topics for that copilot) vs O(topics for all copilots). The downside is that, without fancy data structures, it would likely be at least 16 bytes extra per subscription, and extra work on every subscribe/unsubscribe to maintain it. Even with this, we'd still be doing ~25k unsubscribes per room with 10 copilots.

Maybe easier to discuss in person tomorrow rather than in diff if preferred. Alternatively, we can just commit this and try killing a copilot on the cluster :-)

Test Plan: make check all

Reviewers: #rocketspeed, dhruba

Reviewed By: #rocketspeed, dhruba

Subscribers: dhruba

Differential Revision: https://reviews.facebook.net/D40263