Test for CommandQueue liveness and fix for batch size limit.
Summary: Originally I had a diff for batch size limit on top of the optimisation, but the `if (pending_reads_ == 0) {...}` was a CAS loop. After changing the opmimisation diff to remove the loop, batch size chech and the if got swapped during the rebase. This diff fixes it in an alternative way (we save load(), echange() and fetch_add() on end of read batch).

Test Plan: `make -j all check` + benchmark

Reviewers: #rocketspeed, pja

Reviewed By: #rocketspeed, pja

Subscribers: dhruba

Differential Revision: https://reviews.facebook.net/D41037