Incomplete batch reads issue
Summary: The command queue BatchedRead assumed that either (a) you hit the batch limit, or (b) you read everything. If the user decides for whatever reason to stop reading in the batch at another point, for example, due to flow control constraints, then the write event isn't propagated in the case when we intentionally leave an element in the queue, but not mark it as pending.

Test Plan: make check

Reviewers: stupaq-fb

Reviewed By: stupaq-fb

Subscribers: dhruba

Differential Revision: https://reviews.facebook.net/D42363