Cache data in control tower
Summary:
Cache the tail of the log in the control tower.

The ControlTower has an option to specify the size of the cache
to use. This cache size is shared equally among all the rooms.
The default setting uses a cache size of 100 MB. Specifying a
cache size of 0 disables the cache.

The current implementation does not cache Gaps. It caches only
Message Data.

Test Plan: Unit tests attached.

Reviewers: rocketspeed, pja

Reviewed By: pja

Subscribers: pja, stupaq-fb, dhruba

Differential Revision: https://reviews.facebook.net/D45759