TryWrite for LogDevice queue writes
Summary: Fixes double delete in Control Tower. Cause: storage thread tries to write to topic tailer thread, and Forward returns false. It re-takes ownership of the message, and passes it back to the caller. However, the new Write semantics means that the write actually succeeded and backoff was requested, not retry.

Test Plan: make check all

Reviewers: #rocketspeed, dhruba

Reviewed By: #rocketspeed, dhruba

Subscribers: stupaq-fb, dhruba

Differential Revision: https://reviews.facebook.net/D47217