Allow record callback in storage return false and steal record.
Summary:
This is a clever workaround to have flow control with buffered storage while incurring minimal overhead.

In short.
Any LogStorage implementation should allow backpressure to be applied and LogRecord to be moved away at the same time.
This way buffered storage can claim ownership of the record (which is uniquely owned by either the underlying storage or the buffered storage), yet still apply backpressure to the underlying storage if it fails to deliver part of the buffered record.
In this case, the underlying storage will retry delivery after some time with a record which has the same LogID and Sequencenumber, but empty payload (the paylaod was conceptually stolen in the previous delivery attempt).
In case the record was not stolen, the next delivery attempt will carry unaltered record (with the original payload).

This is necessary to avoid deadlock at the last record and being able to effectively propagate flow control through storage interface.

Test Plan:
  fbconfig -r --clang --sanitize=address rocketspeed/github/src/logdevice/tests rocketspeed/github/src/{tools,server}
  fbmake runtests # I've added a unit test for that

  ./_bin/rocketspeed/github/src/server/rocketspeed --copilot_port 58600 --pilot_port 58600 --tower_port 58500 --logdevice_cluster=logdevice.stupaq.rocketspeed --storage_url=configerator:logdevice/logdevice.stupaq.rocketspeed.conf --logs=1..100 --log_to_stderr --rollcall=false --loglevel=warn
  ./_bin/rocketspeed/github/src/tools/rocketbench/rocketbench --config='one-host;copilot=localhost:58600;pilot=localhost:58600' --start_consumer --delay_subscribe=false --subscribe_rate=1000 --num_topics=10000 --start_producer --num_messages=500000 --message_rate=100000

Reviewers: pja

Reviewed By: pja

Subscribers: lovro, stupaq-fb, dhruba, #rocketspeed

Projects: #rocketspeed

Differential Revision: https://reviews.facebook.net/D48237