DataCache uses bloom filters to remember topics per block.
Summary:
The per-log cache is arranged such that a configurable number
of records comprise a cache block. Blocks maybe configured
to store a bloom filter of topics. The amount of memory
used by blooms can be configured.

The per-block bloom filter is created when a block is
full. A cache lookup for a particular topic(s) may use
the bloom filter to skip over the entire block if that
topic does not exist in that block.

Test Plan: bloom_test

Reviewers: stupaq-fb, pja

Reviewed By: pja

Subscribers: rocketspeed, stupaq-fb, dhruba

Differential Revision: https://reviews.facebook.net/D49443