FindLatestSeqno coalescing
Summary:
When a find time request is already in flight for a log, and another request comes in for the same log then we now re-use the response rather than sending another request. This bounds the number of in-flight requests to the number of logs.

We still need other flow control, but this bounds the memory.

Test Plan:
make check all

Also ran the benchmark with some printfs to ensure the responses were being shared.

Reviewers: #rocketspeed, dhruba, stupaq-fb

Reviewed By: #rocketspeed, dhruba, stupaq-fb

Subscribers: stupaq-fb, dhruba

Differential Revision: https://reviews.facebook.net/D49509