Flow control for FindLatestSeqno
Summary: Limits number of in-flight find time requests. When limit is reached, we buffer them, but remove identical requests for the same log, so the buffer is bounded.

Test Plan:
make check all

Also ran readwrite benchmark with max-inflight == 1, and some printfs to ensure the buffering was happening. Would be great to have a stress test framework that we can run with various options.

Reviewers: stupaq-fb, #rocketspeed

Reviewed By: stupaq-fb, #rocketspeed

Subscribers: stupaq-fb, dhruba

Differential Revision: https://reviews.facebook.net/D49629