Move log state to LogTailer
Summary: This is part 1. Moves the out of order checks into LogTailer so that TopicTailer is guaranteed to receive records in the order it expects, and only for logs that are already opened. The queues from storage -> RS now belongs to LogTailer. More to be done.

Test Plan: make -j check all && fbmake runtests

Reviewers: stupaq-fb, #rocketspeed

Reviewed By: stupaq-fb, #rocketspeed

Subscribers: stupaq-fb, dhruba

Differential Revision: https://reviews.facebook.net/D50433