Cache re-entry flow control
Summary:
This handles the case when a log reader is opened, reads some logs, discovers that there are some records in the cache, starts delivering those, but then hits a full queue. Previously, the reader would just restart reading from the log at this point (not fully utilizing the cache).

This diff now has a separate source for cache readers. When reading from the cache, we stop the log reader, and only start it again once we finish reading the cache. Normal flow control mechanics handle backpressure while reading the cache.

This implementation is suboptimal because there is a single source for all readers (per thread). Ideally there would be a source per reader/per log, but that would likely introduce too many FDs. This is a fairly simple change once we have the merged FD that Mateusz is working on.

Test Plan:
make check all

Also ran Test3 with room->client queue size of 10.

Reviewers: #rocketspeed, stupaq-fb

Reviewed By: #rocketspeed, stupaq-fb

Subscribers: stupaq-fb, dhruba

Differential Revision: https://reviews.facebook.net/D50685