Cache assertion fix
Summary: ProcessRecord had a check to ensure that next seqno was the one expected. When iterating a filtered cache however, that might not be the case. It's still the case when reading from logdevice, so I moved the check there, and weakened the check for cache reads.

Test Plan: make -j check all && Test3

Reviewers: #rocketspeed, stupaq-fb

Reviewed By: #rocketspeed, stupaq-fb

Subscribers: stupaq-fb, dhruba

Differential Revision: https://reviews.facebook.net/D51519