Fixed port::Eventfd on OSX.
Summary:
We require the following two invariants to be preserved by any implementation of the port::Eventfd (be it the pipe of eventfd based one):
* sum of all written values = sum of all read values + current value,
* if current value > 0 then eventfd is readable.

The first invariant states that the value of the counter does not change spontaneously, while the second ensures that if there is a notification to be read, libevent will invoke appropriate callback.
The problem here is that without this patch writing 0 would render the fd readable, which might render the following code to loop forever:

  while (fd is readable) {
    eventfd_t x;
    fd.read_event(&x);
    --x;
    fd.write_event(x);
  }

See https://reviews.facebook.net/D49761 for more context.

Test Plan: `make -j all check`

Reviewers: #rocketspeed, dyniusz

Reviewed By: #rocketspeed, dyniusz

Subscribers: dyniusz, dhruba

Differential Revision: https://reviews.facebook.net/D51753