[WIP] change control_command_queue_ to UnboundedMPSCCommandQueue
Summary:
This is a work-in-progress diff. There might be code redundancy in diff, so call for comment.

control_command_queue_ is concurrently implemented using class CommandQueue, which is a Fixed-size, single-producer, single-consumer queue.
We want it to be unbounded, to ensure that control commands never fail.

Test Plan:
make -j all check
https://phabricator.fb.com/P56097576

Reviewers: pja, #rocketspeed

Reviewed By: pja, #rocketspeed

Differential Revision: https://reviews.facebook.net/D53115