Bring back storage_test.cc
Summary: I removed it in a previous diff because it was included by the
new Makefile rules for tests and I thought it was an old test that
didn't work any more. Didn't realise it was still used in fbcode.
I've filtered it out from the Makefile rule.

Test Plan: make -j check all

Reviewers:

CC:

Task ID: #

Blame Rev: