Batch Delivery
Summary: Extend the RocketSpeed communication protocol to support sending multiple MessageDeliver (on different subscriptions) in a single message. This is a new message type. The purpose of this message type is to allow consumers to process a set of messages as a batch for efficiency, but could also be used later to communicate atomicity of a set of messages.

Test Plan:
 make -j all check

Added new MessageDeliverBatch test and updated StackRocketeerTest.

Reviewers: stupaq-fb, #rocketspeed, pja

Reviewed By: #rocketspeed, pja

Subscribers: dhruba, stupaq-fb

Differential Revision: https://reviews.facebook.net/D53853