Subscriber interface.
Summary:
Also removed the "correlated ticks" of all per-shard subscribers. Not sure if this might ever be a problem, but we could solve it at the EventLoop level (some semi-inteligent smearing would be nice).

Also got rid of MultiShardSubscriber::Start() that was called on wrong thread... It was probably fine as we never attached MultiShardSubscriber to a running loop, but it's generally not okay, as we use thread-unsafe libevent.

Test Plan: `make -j all check`

Reviewers: pja, #rocketspeed

Reviewed By: pja, #rocketspeed

Subscribers: dhruba

Differential Revision: https://reviews.facebook.net/D54159