Fix rate limiter race
Summary:
Fix rate limiting and test.

The basic issue was this: RateLimiter allows X events every T seconds. If you perform X events and then ask again 0.999T later, it will deny you. RateLimiterSink sets up a timer that fires every T seconds. As these timers are different, they are subject to races, so it might be the case that the RateLimiterSink wakes up 0.999T later, gets denied, and then goes back to sleep for T seconds, effectively halving the throughput.

The fix is to just make the RateLimiterSink wake up and check 10x more frequently to mitigate the effect of the race. There was also an off-by-one in the RateLimiter usage. I also changed the test to be less prescriptive about its requirements.

Test Plan: make -j check all

Reviewers: saf-fb, #rocketspeed

Reviewed By: saf-fb, #rocketspeed

Subscribers: saf-fb

Differential Revision: https://reviews.facebook.net/D56145