Fix build configuration issues on CentOS 7
Summary:
In the Makefile, ".PRECOUS %test: <...>" causes make to throw an error and not do anything.  Separated into two rules.
In the Facebook platform detection section, the rule was CentOS 6, or anything else, which caused the CentOS 5 build chain to be used for CentOS 7.  It has many errors on CentOS 7, so I simply switched the rule to be 6 or 7.  The CentOS 6 build chain works fine on CentOS 7.

Test Plan: Have a Facebook CentOS 7 build machine.  Follow the usual build instructions.

Reviewers: stupaq-fb, #rocketspeed

Reviewed By: stupaq-fb, #rocketspeed

Differential Revision: https://reviews.facebook.net/D58113