Demultiplex test for TailCollapsingSubscriber
Summary:
This diff is a test of demultiplexing in TailCollapsingSubscriber.  The task is to improve memory usage when dropping back to a single subscriber.
A separate diff will have that change, this verifies that it works (this test passes with or without the change)

Test Plan: Run client_test.test

Reviewers: stupaq-fb, #rocketspeed

Reviewed By: stupaq-fb, #rocketspeed

Differential Revision: https://reviews.facebook.net/D58461